+++ packages/app/vite-utils/vite.base.config.ts
@@ -2,11 +2,12 @@ import react from '@vitejs/plugin-react';
 import path from 'node:path';
 import type { PluginOption, UserConfig } from 'vite';
 import cleanPlugin from 'vite-plugin-clean';
+import { plugin as viteMdPlugin, Mode } from 'vite-plugin-markdown';
 import tsconfigPaths from 'vite-tsconfig-paths';
 
 import '../load.envs.js';
 
-const linkDeps = process.env.LINK_DEPS?.trim().split(' ') || [];
+const linkDeps = process.env.LINK_DEPS?.trim().split(' ').filter(Boolean) || [];
 
 // https://vitejs.dev/config/
 const baseConfig: UserConfig = {
@@ -38,6 +39,9 @@ const baseConfig: UserConfig = {
   },
   plugins: [
     react(),
+    viteMdPlugin({
+      mode: [Mode.REACT],
+    }),
     tsconfigPaths(),
     {
       ...cleanPlugin({
@@ -57,14 +61,17 @@ const baseConfig: UserConfig = {
   define: {
     'process.env': {},
   },
-  ...(process.env.WITH_PNPM_LINKS && {
+  ...(!!linkDeps.length && {
     resolve: {
       alias: linkDeps.reduce((obj, dep) => {
-        // eslint-disable-next-line no-param-reassign
-        obj[dep] = path.resolve(
-          __dirname,
-          `../node_modules/${dep}/dist/index.mjs`
-        );
+        // remove TS SDK as it's not needed to resolve alias anymore.
+        if (!/^fuels?|@fuel-ts/.test(dep)) {
+          // eslint-disable-next-line no-param-reassign
+          obj[dep] = path.resolve(
+            __dirname,
+            `../node_modules/${dep}/dist/index.mjs`
+          );
+        }
         return obj;
       }, {}),
     },