+++ packages/asset-list/src/assets/index.ts
@@ -1,0 +1,1 @@
+export { default as ETH } from './eth.svg';