+++ packages/asset-list/src/index.ts
@@ -1,9 +1,11 @@
+import * as assets from './assets';
+
 export const AssetList = [
   {
     assetId:
       '0x0000000000000000000000000000000000000000000000000000000000000000',
     name: 'Ethereum',
     symbol: 'ETH',
-    imageUrl: `assets/eth.svg`,
+    imageUrl: String(assets.ETH),
   },
 ];