+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @fuel-wallet/types
 
+## 0.8.0
+
 ## 0.7.2
 
 ### Patch Changes